Keep original attribute values in decorated functions
When using attributes inside decorated functions, and the attribute
was redefined later on in the same module, the last known value was
used when the decorated function was redefined, leading to strainge
bugs.

Instead, the value that the attribute had on the position of the
original definition of the function should be used.

This commit fixes that, remembering the attribute values on the
moment of the function definition, and re-emitting those module
attributes later when the decorated function is emitted.